+++ tests/model.rs
@@ -204,3 +204,22 @@ fn test_universal3d() {
     let fmt = FileFormat::from_file("fixtures/model/sample.u3d").unwrap();
     assert_eq!(fmt, FileFormat::Universal3d);
 }
+
+#[test]
+fn test_universal_scene_description_ascii() {
+    let fmt = FileFormat::from_file("fixtures/model/sample.usda").unwrap();
+    assert_eq!(fmt, FileFormat::UniversalSceneDescriptionAscii);
+}
+
+#[test]
+fn test_universal_scene_description_binary() {
+    let fmt = FileFormat::from_file("fixtures/model/sample.usdc").unwrap();
+    assert_eq!(fmt, FileFormat::UniversalSceneDescriptionBinary);
+}
+
+#[cfg(feature = "reader-zip")]
+#[test]
+fn test_universal_scene_description_zipped() {
+    let fmt = FileFormat::from_file("fixtures/model/sample.usdz").unwrap();
+    assert_eq!(fmt, FileFormat::UniversalSceneDescriptionZipped);
+}